fetch chat list from view model
