+++ app/src/main/java/com/mobileforce/hometeach/ui/message/ChatListFragment.kt
@@ -6,8 +6,6 @@ import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
 import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.LinearLayoutManager
-import com.google.firebase.firestore.ktx.firestore
-import com.google.firebase.ktx.Firebase
 import com.mobileforce.hometeach.R
 import com.mobileforce.hometeach.adapters.RecyclerViewAdapter
 import com.mobileforce.hometeach.adapters.ViewHolder
@@ -24,7 +22,6 @@ class ChatListFragment :
     Fragment(R.layout.fragment_chat_list)/* the constructor takes the layout which will be inflated at onCreateView*/ {
 
 
-    private val db = Firebase.firestore
     private val viewModel: ChatViewModel by sharedViewModel()
 
     private val chatListItemClickListener: (Chat) -> Unit = { chat ->
@@ -61,55 +58,18 @@ class ChatListFragment :
             layoutManager = LinearLayoutManager(requireContext())
         }
 
+        viewModel.fetchUserChatList()
 
         viewModel.user.observe(viewLifecycleOwner, Observer { user ->
-
-            db.clearPersistence()
             user?.let {
-                db
-                    .collection("user")
-                    .document(it.id)
-                    .collection("connect")
-                    .addSnapshotListener { snapshot, error ->
 
-                        error?.let {
-                            return@addSnapshotListener
-                        }
+            }
+        })
 
-                        if (snapshot != null) {
+        viewModel.chatList.observe(viewLifecycleOwner, Observer {
 
-                            val chatList = mutableListOf<Chat>()
-                            for (doc in snapshot.documents) {
-
-                                val name = doc.getString("full_name")
-                                val senderId = doc.getString("id")
-                                val connectId = doc.getString("connect_id")
-                                val lastMessage = doc.getString("last_message")
-
-
-                                val chat = Chat(
-                                    chatId = connectId.toString(),
-                                    senderName = name.toString(),
-                                    senderPhoto = "",
-                                    lastMessage = lastMessage.toString(),
-                                    lastMessageTime = "Today",
-                                    senderId = senderId!!
-
-                                )
-
-                                chatList.add(chat)
-
-                            }
-
-                            chatListAdpter.submitList(chatList)
-
-
-                        } else {
-                            println("Current data: null")
-                        }
-
-
-                    }
+            it?.let { list ->
+                chatListAdpter.submitList(list)
             }
 
         })
